+++ CHANGELOG.md
@@ -27,3 +27,7 @@
 * Change get_it version from 3.1.0 to 4.0.0;
 * Change google_maps_flutter version from 0.5.22 to 0.5.24+1;
 * Change flutter spinkit version from 4.1.1+1 to 4.1.2.
+
+## [0.1.4] - Add properties on AppBar parser
+
+* Add many properties on AppBarSchemaWidgetParser.